Implement HITANv6 in the TAN namespace, add a TanRequest interface
