+++ lib/Fhp/Segment/HITANS/HITANSv6.php
@@ -10,7 +10,10 @@ use Fhp\Segment\BaseGeschaeftsvorfallparameter;
  * Bezugssegment: HKVVB
  * Sender: Kreditinstitut
  *
- * @link: https://www.hbci-zka.de/dokumente/spezifikation_deutsch/archiv/FinTS_V3.0_2017-10-06-FV_RM.zip
+ * @link: https://www.hbci-zka.de/dokumente/spezifikation_deutsch/fintsv3/FinTS_3.0_Security_Sicherheitsverfahren_PINTAN_2018-02-23_final_version.pdf
+ * Section: B.5.1 c)
+ *
+ * TODO Move HITANS to the TAN namespace. Probably we can drop all HITANS older than v6.
  */
 class HITANSv6 extends BaseGeschaeftsvorfallparameter implements HITANS
 {
