+++ lib/Fhp/Segment/HITANS/VerfahrensparameterZweiSchrittVerfahrenV6.php
@@ -6,9 +6,11 @@ use Fhp\Segment\BaseDeg;
 
 class VerfahrensparameterZweiSchrittVerfahrenV6 extends BaseDeg implements VerfahrensparameterZweiSchrittVerfahren
 {
+    const PROZESSVARIANTE_2 = 2;
+
     /** @var int Allowed values: 900 through 997 */
     public $sicherheitsfunktion;
-    /** @var int Allowed values: 1, 2; See specification for details */
+    /** @var int Allowed values: 1, 2; See specification or {@link HKTANv6#$tanProzess} for details. */
     public $tanProzess;
     /** @var string */
     public $technischeIdentifikationTanVerfahren;
@@ -29,6 +31,8 @@ class VerfahrensparameterZweiSchrittVerfahrenV6 extends BaseDeg implements Verfa
     /** @var bool */
     public $mehrfachTanErlaubt;
     /**
+     * In case of multi-TAN (see {@link #$mehrfachTanErlaubt}), this specifies whether all TANs must be entered in the
+     * same dialog and at the same time, or not.
      * 1 TAN nicht zeitversetzt / dialogübergreifend erlaubt
      * 2 TAN zeitversetzt / dialogübergreifend erlaubt
      * 3 beide Verfahren unterstützt
